code review refactor
Moved useExpandable and the toggle into a shared study-code-collapse.tsx that both views import, so they can't drift again. That also resolves the caret mismatch — the shared toggle uses rotate(-90deg), so both pages now match. Behavior unchanged (same test-id/labels/aria-expanded).
